on repars sur l,ancien systeme

+++ sources/draw/draw_game.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   draw_game.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cllovio <cllovio@student.42lyon.fr>        +#+  +:+       +#+        */
+/*   By: cllovio <cllovio@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/09 14:20:24 by cllovio           #+#    #+#             */
-/*   Updated: 2023/10/12 09:32:07 by cllovio          ###   ########.fr       */
+/*   Updated: 2023/10/13 15:28:04 by cllovio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,14 +47,14 @@ static void	draw_wall(int *y, t_raycasting *raycasting, int y_max, long size_wal
 	{
 		if (raycasting->data->horizontal == true)
 		{
-			x_pixel_get = ((raycasting->data->collision_cor[1] % SQUARE_SIZE) * 100) / SQUARE_SIZE;
-			y_pixel_get = ((*y - ((SIZE_Y >> 1) - (size_wall >> 1))) * 100) / (size_wall);
+			x_pixel_get = (int)((((int)raycasting->data->collision_cor[1] % SQUARE_SIZE) * 100) / SQUARE_SIZE);
+			y_pixel_get = (int)(((*y - ((SIZE_Y >> 1) - (size_wall >> 1))) * 100) / (size_wall));
 			color = my_mlx_pixel_get(&raycasting->data->N_texture, x_pixel_get, y_pixel_get);
 		}
 		else if (raycasting->data->horizontal == false)
 		{
-			x_pixel_get = ((raycasting->data->collision_cor[0] % SQUARE_SIZE) * 100) / SQUARE_SIZE;
-			y_pixel_get = ((*y - ((SIZE_Y >> 1) - (size_wall >> 1))) * 100) / (size_wall);			
+			x_pixel_get = ((((int)raycasting->data->collision_cor[0] % SQUARE_SIZE) * 100) / SQUARE_SIZE);
+			y_pixel_get = (int)(((*y - ((SIZE_Y >> 1) - (size_wall >> 1))) * 100) / (size_wall));			
 			color = my_mlx_pixel_get(&raycasting->data->N_texture, x_pixel_get, y_pixel_get);
 		}
 		my_mlx_pixel_put(raycasting->draw, raycasting->x, *y, color);
